use better html/css
