+++ src/app/inputForm.js
@@ -6,8 +6,9 @@ import {
 } from '../common/calculation';
 import { raceOptions } from '../common/config';
 import {
+  disableUnitInput,
   displayResult,
-  handleUnitInputChange,
+  enableUnitInput,
   setErrorLabel,
 } from '../common/domManipulation';
 
@@ -33,35 +34,26 @@ export function initializeInput() {
 
   inputRace.onchange = function() {
     setErrorLabel(false);
-    const unitInputGroup = document.getElementsByClassName('input-unit');
-    if (unitInputGroup && unitInputGroup.length > 0) {
-      enableUnitInput(unitInputGroup);
-    }
+    enableUnitInput();
   };
 
   inputRace.onblur = function(e) {
     const value = e.target.value;
-    const unitInputGroup = document.getElementsByClassName('input-unit');
-    if (!(unitInputGroup && unitInputGroup.length > 0)) {
-      return;
-    }
     if (
       raceOptions.some(function(race) {
         return race.name === value;
       })
     ) {
-      if (!unitInputGroup[0].classList.contains('disabled')) {
-        disableUnitInput(unitInputGroup);
-      }
+      disableUnitInput();
     } else {
-      if (unitInputGroup[0].classList.contains('disabled')) {
-        enableUnitInput(unitInputGroup);
-      }
+      enableUnitInput();
     }
   };
 
   inputUnit.forEach(function(unitInput) {
-    unitInput.onchange = handleUnitInputChange(unitInput);
+    unitInput.onchange = function() {
+      setErrorLabel(false);
+    };
   });
 }
 
@@ -119,28 +111,4 @@ function getAndValidateInput() {
     unit,
     time: inputTime.value,
   };
-}
-
-export function disableUnitInput(unitInputGroup) {
-  const selectedUnit = document.querySelector('input[name=unit]:checked');
-  if (selectedUnit) {
-    selectedUnit.setAttribute('checked', false);
-    const selectedUnitLabel = unitInputGroup[0].querySelector('label.checked');
-    if (selectedUnitLabel) {
-      selectedUnitLabel.classList.remove('checked');
-    }
-  }
-  const allUnitInputs = document.querySelectorAll('input[name=unit]');
-  allUnitInputs.forEach(function(unitInput) {
-    unitInput.setAttribute('disabled', 'disabled');
-  });
-  unitInputGroup[0].classList.add('disabled');
-}
-
-function enableUnitInput(unitInputGroup) {
-  const allUnitInputs = document.querySelectorAll('input[name=unit]');
-  allUnitInputs.forEach(function(unitInput) {
-    unitInput.removeAttribute('disabled');
-  });
-  unitInputGroup[0].classList.remove('disabled');
 }
