+++ src/app/styles.scss
@@ -65,30 +65,10 @@ h2 {
     justify-content: space-between;
     align-items: stretch;
 
-    &.disabled {
-      opacity: 0.3;
-      pointer-events: none;
-    }
-
     .input-container {
       display: flex;
-      justify-content: center;
-      align-items: center;
       position: relative;
       margin: 0 8px;
-      background-color: white;
-      border-radius: $default-border-radius;
-      border: 1px solid $dark-grey;
-      cursor: pointer;
-
-      &:hover,
-      &:focus {
-        background-color: $light-grey;
-      }
-
-      &.checked {
-        background-color: $light-yellow;
-      }
 
       &:first-child {
         margin-left: 0;
@@ -98,13 +78,46 @@ h2 {
         margin-right: 0;
       }
 
+      label {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        background-color: white;
+        border-radius: $default-border-radius;
+        border: 1px solid $dark-grey;
+        cursor: pointer;
+        height: 100%;
+        width: 100%;
+
+        &:hover {
+          background-color: $light-grey;
+        }
+      }
+
       input[type='radio'] {
+        -moz-appearance: none;
+        -webkit-appearance: none;
+        appearance: none;
+        border: none;
         position: absolute;
         height: 0;
         width: 0;
         top: 0;
         left: 0;
         z-index: -10;
+
+        &:checked + label {
+          background-color: $light-yellow;
+        }
+
+        &[disabled] + label {
+          opacity: 0.3;
+          pointer-events: none;
+        }
+
+        &:focus + label {
+          outline: rgba(0, 103, 244, 0.247) auto 5px;
+        }
       }
     }
   }
